feature: added functionality to arrows
