+++ lib/widgets/cloud_carousel.dart
@@ -73,20 +73,3 @@ final List<Widget> slidingImages = imgList
       ),
     )
     .toList();
-
-class CloudCarousel extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: CarouselSlider(
-        options: CarouselOptions(
-          height: MediaQuery.of(context).size.height * 0.21,
-          autoPlay: true,
-          aspectRatio: 16 / 9,
-          enlargeCenterPage: true,
-        ),
-        items: slidingImages,
-      ),
-    );
-  }
-}
